Working EDMX Scaffolding
