+++ src/GUI/EFCorePowerTools.sln
@@ -23,7 +23,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "efpt30.core", "efpt30.core\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RevEng.Shared", "RevEng.Shared\RevEng.Shared.csproj", "{D4594A0C-B02F-469D-9521-F0BDD5A30FC0}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "efreveng", "efreveng\efreveng.csproj", "{EBD761ED-96D2-44F6-9099-9D376CE7181F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "efreveng", "efreveng\efreveng.csproj", "{EBD761ED-96D2-44F6-9099-9D376CE7181F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DacFxStrongTypedCore", "DacFxStrongTypedCore\DacFxStrongTypedCore.csproj", "{A9D41FE5-9FB7-4640-ADFF-217BA39F1CE3}"
 EndProject
@@ -47,7 +47,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RevEng.Core.Abstractions", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ErikEJ.EntityFrameworkCore.5.Edmx", "ErikEJ.EntityFrameworkCore.5.Edmx\ErikEJ.EntityFrameworkCore.5.Edmx.csproj", "{1CDE7157-E394-495D-B61E-2424C5FBE6DC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ErikJ.EntityFrameworkCore.Edmx", "ErikJ.EntityFrameworkCore.Edmx\ErikJ.EntityFrameworkCore.Edmx.csproj", "{1954049E-DA77-4903-AF07-29D735DA8540}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ErikEJ.EntityFrameworkCore.Edmx", "ErikEJ.EntityFrameworkCore.Edmx\ErikEJ.EntityFrameworkCore.Edmx.csproj", "{1954049E-DA77-4903-AF07-29D735DA8540}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ErikEJ.EntityFrameworkCore.5.DgmlBuilder", "ErikEJ.EntityFrameworkCore.5.DgmlBuilder\ErikEJ.EntityFrameworkCore.5.DgmlBuilder.csproj", "{9212E4DF-87A4-40B3-A1E7-858C0A0F4DC3}"
 EndProject
