+++ src/GUI/ErikEJ.EntityFrameworkCore.Edmx/ErikEJ.EntityFrameworkCore.Edmx.csproj
