+++ src/GUI/ErikEJ.EntityFrameworkCore.Edmx/Scaffolding/EdmxDatabaseModelFactory.cs
@@ -3,17 +3,23 @@ using System.Collections.Generic;
 using System.Data.Common;
 using System.IO;
 using System.Linq;
+using System.Xml;
 using System.Xml.Linq;
+using System.Xml.XPath;
 using LinqToEdmx;
+using LinqToEdmx.MapV3;
+using LinqToEdmx.Model.ConceptualV3;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Diagnostics;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Scaffolding;
 using Microsoft.EntityFrameworkCore.Scaffolding.Metadata;
 
-namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
+namespace ErikEJ.EntityFrameworkCore.Edmx.Scaffolding
 {
-    // This is the conceptual part of the model
-    public class EdmxRelationalModelFactory : IDatabaseModelFactory
+    // This is the storage side of the scaffolding logic
+    // Because the EDMX format was not designed with the database creation in mind, the SSDL part of the schema won't contains default values, extended properties, primary keys, unique constraints and index informations.
+    public class EdmxDatabaseModelFactory : IDatabaseModelFactory
     {
         // SQL Server db types
         private static readonly ISet<string> SQLServerDateTimePrecisionTypes = new HashSet<string> { "datetimeoffset", "datetime2", "time" };
@@ -21,85 +27,15 @@ namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
         private static readonly ISet<string> SQLServerMaxLengthRequiredTypes
             = new HashSet<string> { "binary", "varbinary", "char", "varchar", "nchar", "nvarchar" };
 
+        // FIXME Add PostgreSQL db types
+
+        // FIXME Add MySql db types
+
         private readonly IDiagnosticsLogger<DbLoggerCategory.Scaffolding> _logger;
 
-        public DatabaseModel Create(string edmxPath, DatabaseModelFactoryOptions options)
+        public EdmxDatabaseModelFactory(IDiagnosticsLogger<DbLoggerCategory.Scaffolding> logger)
         {
-            if (string.IsNullOrEmpty(edmxPath))
-            {
-                throw new ArgumentException(@"invalid path", nameof(edmxPath));
-            }
-            if (!File.Exists(edmxPath))
-            {
-                throw new ArgumentException($"Edmx file not found: {edmxPath}");
-            }
-
-            var schemas = options.Schemas;
-            var tables = options.Tables;
-
-            var dbModel = new DatabaseModel
-            {
-                DatabaseName = Path.GetFileNameWithoutExtension(edmxPath),
-                DefaultSchema = schemas.Count() > 0 ? schemas.First() : "dbo"
-            };
-
-            // Detect the EDMX file version
-            // FIXME Provide an alternative to not load the whole document in memory /!\
-            XDocument edmxFile = XDocument.Load(edmxPath);
-
-            var edmxVersion = ((XElement)edmxFile.FirstNode).FirstAttribute.Value;
-
-            // FIXME Assume EDMX version is 3.0 for now
-
-            if (string.Compare(edmxVersion, @"3.0", StringComparison.InvariantCultureIgnoreCase) == 0)
-            {
-                var edmxv3 = LinqToEdmx.EdmxV3.Load(edmxPath);
-
-                // The conceptual entity name can be anything.
-                // Most of the time it is the pluralized srore table name though.
-                var entitySets = edmxv3.GetItems<LinqToEdmx.Model.ConceptualV3.EntityContainer>().First().EntitySets;
-
-                foreach (var entity in entitySets)
-                {
-                    //var mappings = edmxv3.GetItems<LinqToEdmx.MapV3.EntitySetMapping>().First(esm => esm.Name == entity.Name).Name;
-                    // We should go through the mapping to find the the store name of a conceptual entity
-                    var storeName = edmxv3.GetItems<LinqToEdmx.MapV3.EntitySetMapping>().First(esm => esm.Name == entity.Name).EntityTypeMappings.First().MappingFragments.First().StoreEntitySet;
-                    var item = edmxv3.GetItems<LinqToEdmx.Model.ConceptualV3.EntityType>().First(e => e.Name == entity.EntityType.Replace("Self.", string.Empty));
-
-                    var dbTable = new DatabaseTable
-                    {
-                        Name = entity.Name,
-                        Schema = GetSchemaNameV3(edmxv3, storeName),
-                    };
-
-
-                    GetColumnsV3(item, dbTable/*, typeAliases, model.GetObjects<TSqlDefaultConstraint>(DacQueryScopes.UserDefined).ToList()*/, edmxv3);
-                    GetPrimaryKeyV3(item, dbTable);
-
-                    dbModel.Tables.Add(dbTable);
-                }
-
-                foreach (var entity in entitySets)
-                {
-                    // We should go through the mapping to find the the store name of a conceptual entity
-                    var storeName = edmxv3.GetItems<LinqToEdmx.MapV3.EntitySetMapping>().First(esm => esm.Name == entity.Name).EntityTypeMappings.First().MappingFragments.First().StoreEntitySet;
-
-                    GetForeignKeysV3(edmxv3, entity.Name, storeName, dbModel);
-
-                    // Unique constraints are not available in the model
-                    // GetUniqueConstraints(item, dbModel);
-
-                    // Indexes are not available in the model
-                    // GetIndexes(item, dbModel);
-                }
-            }
-
-            return dbModel;
-        }
-
-        public DatabaseModel Create(DbConnection connection, DatabaseModelFactoryOptions options)
-        {
-            throw new NotImplementedException();
+            _logger = logger;
         }
 
         // FIXME There could be a table or a view with the same name in multiple schema /!\
@@ -135,15 +71,81 @@ namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
             throw new InvalidOperationException(@$"Unable to identify the database object schema for entity [{entityName}]. This usually indicates a bug");
         }
 
-        private void GetColumnsV3(LinqToEdmx.Model.ConceptualV3.EntityType item, DatabaseTable dbTable/*, IReadOnlyDictionary<string, (string storeType, string typeName)> typeAliases, List<TSqlDefaultConstraint> defaultConstraints,*/, LinqToEdmx.EdmxV3 model)
+        public DatabaseModel Create(string edmxPath, DatabaseModelFactoryOptions options)
+        {
+            if (string.IsNullOrEmpty(edmxPath))
+            {
+                throw new ArgumentException(@"invalid path", nameof(edmxPath));
+            }
+            if (!File.Exists(edmxPath))
+            {
+                throw new ArgumentException($"Edmx file not found: {edmxPath}");
+            }
+
+            var schemas = options.Schemas;
+            var tables = options.Tables;
+
+            var dbModel = new DatabaseModel
+            {
+                DatabaseName = Path.GetFileNameWithoutExtension(edmxPath),
+                DefaultSchema = schemas.Count() > 0 ? schemas.First() : "dbo"
+            };
+
+            // Detect the EDMX file version
+            // FIXME Provide an alternative to not load the whole document in memory /!\
+            XDocument edmxFile = XDocument.Load(edmxPath);
+
+            var edmxVersion = ((XElement)edmxFile.FirstNode).FirstAttribute.Value;
+
+            // FIXME Assume EDMX version is 3.0 for now
+
+            if (string.Compare(edmxVersion, @"3.0", StringComparison.InvariantCultureIgnoreCase) == 0)
+            {
+                var edmxv3 = LinqToEdmx.EdmxV3.Load(edmxPath);
+                var items = edmxv3.GetItems<LinqToEdmx.Model.StorageV3.EntityTypeStore>();
+
+                foreach (var item in items)
+                {
+                    var dbTable = new DatabaseTable
+                    {
+                        Name = item.Name,
+                        Schema = GetSchemaNameV3(edmxv3, item.Name),
+                    };
+
+                    GetColumnsV3(item, dbTable/*, typeAliases, model.GetObjects<TSqlDefaultConstraint>(DacQueryScopes.UserDefined).ToList()*/, edmxv3);
+                    GetPrimaryKeyV3(item, dbTable);
+
+                    dbModel.Tables.Add(dbTable);
+                }
+
+                foreach (var item in items)
+                {
+                    GetForeignKeysV3(edmxv3, item, dbModel);
+                    // Unique constraints are not available in the model
+                    // GetUniqueConstraints(item, dbModel);
+
+                    // Indexes are not available in the model
+                    // GetIndexes(item, dbModel);
+                }
+            }
+
+            return dbModel;
+        }
+
+        public DatabaseModel Create(DbConnection connection, DatabaseModelFactoryOptions options)
+        {
+            throw new NotImplementedException();
+        }
+
+        private void GetColumnsV3(LinqToEdmx.Model.StorageV3.EntityTypeStore item, DatabaseTable dbTable/*, IReadOnlyDictionary<string, (string storeType, string typeName)> typeAliases, List<TSqlDefaultConstraint> defaultConstraints,*/, LinqToEdmx.EdmxV3 model)
         {
             var tableColumns = item.Properties;
-            //.Where(i => !i.GetProperty<bool>(Column.IsHidden)
-            //&& i.ColumnType != ColumnType.ColumnSet
-            //// Computed columns not supported for now
-            //// Probably not possible: https://stackoverflow.com/questions/27259640/get-datatype-of-computed-column-from-dacpac
-            //&& i.ColumnType != ColumnType.ComputedColumn
-            //);
+                //.Where(i => !i.GetProperty<bool>(Column.IsHidden)
+                //&& i.ColumnType != ColumnType.ColumnSet
+                //// Computed columns not supported for now
+                //// Probably not possible: https://stackoverflow.com/questions/27259640/get-datatype-of-computed-column-from-dacpac
+                //&& i.ColumnType != ColumnType.ComputedColumn
+                //);
 
             foreach (var col in tableColumns)
             {
@@ -170,7 +172,6 @@ namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
                     //        ? ValueGenerated.OnAddOrUpdate
                     //        : default(ValueGenerated?)
                 };
-
                 if (storeType == "rowversion")
                 {
                     dbColumn["ConcurrencyToken"] = true;
@@ -193,7 +194,7 @@ namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
         }
 
         // FIXME Add provider string to deal with provider specific data type string to be returned. Only SQL Server for now.
-        private string GetStoreTypeV3(LinqToEdmx.Model.ConceptualV3.EntityProperty col)
+        private string GetStoreTypeV3(LinqToEdmx.Model.StorageV3.EntityProperty col)
         {
             // SQL Server
             string dataTypeName = col.Type.ToString();
@@ -234,11 +235,11 @@ namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
             // FIXME MySQL
         }
 
-        private void GetPrimaryKeyV3(LinqToEdmx.Model.ConceptualV3.EntityType conceptualTable, DatabaseTable dbTable)
+        private void GetPrimaryKeyV3(LinqToEdmx.Model.StorageV3.EntityTypeStore table, DatabaseTable dbTable)
         {
-            if (conceptualTable.Key.PropertyRefs.Count() == 0) return;
+            if (table.Key.PropertyRefs.Count() == 0) return;
 
-            var pk = conceptualTable.Key;
+            var pk = table.Key;
             var primaryKey = new DatabasePrimaryKey
             {
                 // We do not have information about the primary key name in the model.
@@ -253,7 +254,7 @@ namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
             //    primaryKey["SqlServer:Clustered"] = false;
             //}
 
-            foreach (var pkCol in conceptualTable.Key.PropertyRefs)
+            foreach (var pkCol in table.Key.PropertyRefs)
             {
                 var dbCol = dbTable.Columns
                     .Single(c => c.Name == pkCol.Name);
@@ -265,20 +266,20 @@ namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
 
         }
 
-        private void GetForeignKeysV3(EdmxV3 model, string conceptualTableName, string storeTableName, DatabaseModel dbModel)
+        private void GetForeignKeysV3(EdmxV3 model, LinqToEdmx.Model.StorageV3.EntityTypeStore storeTable, DatabaseModel dbModel)
         {
             var table = dbModel.Tables
-                .Single(t => t.Name == conceptualTableName
-                && t.Schema == GetSchemaNameV3(model, storeTableName));
+                .Single(t => t.Name == storeTable.Name
+                && t.Schema == GetSchemaNameV3(model, storeTable.Name));
 
             // The foreign key informations are stored in the Association Object
-            var associations = model.GetItems<LinqToEdmx.Model.StorageV3.Association>().Where(a => a.ReferentialConstraint.Dependent.Role == conceptualTableName);
+            var associations = model.GetItems<LinqToEdmx.Model.StorageV3.Association>().Where(a => a.ReferentialConstraint.Dependent.Role == storeTable.Name);
 
             if (!associations.Any())
             {
                 // Give a chance to a reflexive constraint
                 // TODO make sure that this is the right way to deal with this case.
-                var endType = string.Concat(@"Self.", conceptualTableName);
+                var endType = string.Concat(@"Self.", storeTable.Name);
                 associations = model.GetItems<LinqToEdmx.Model.StorageV3.Association>().Where(a => a.Ends.All(e => e.Type == endType));
             }
 
@@ -298,7 +299,7 @@ namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
                 // Look for the table to which the columns are constrained
                 // Remember that we could be constrained to ourself :)
                 var principalTable = dbModel.Tables.SingleOrDefault(t => t.Name == entityName && t.Schema == GetSchemaNameV3(model, entityName));
-                if (principalTable == null)
+                if ( principalTable == null)
                 {
                     entityName = associationSet.Ends.Single(e => e.Role == entityName).EntitySet;
                     principalTable = dbModel.Tables.SingleOrDefault(t => t.Name == entityName && t.Schema == GetSchemaNameV3(model, entityName));
@@ -316,7 +317,7 @@ namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
                     Table = table,
                     // The table to which the columns are constrained (Principal in the Edmx)
                     PrincipalTable = principalTable,
-
+                    
                 };
 
                 var end = associationSet.Ends[0];
@@ -351,6 +352,27 @@ namespace ErikJ.EntityFrameworkCore.Edmx.Scaffolding
                 {
                     table.ForeignKeys.Add(foreignKey);
                 }
+            }
+        }
+
+        private static ReferentialAction? ConvertToReferentialAction(OnAction onDeleteAction)
+        {
+            switch (onDeleteAction.Action)
+            {
+                case "NoAction":
+                    return ReferentialAction.NoAction;
+
+                case "Cascade":
+                    return ReferentialAction.Cascade;
+
+                case "SetNull":
+                    return ReferentialAction.SetNull;
+
+                case "SetDefault":
+                    return ReferentialAction.SetDefault;
+
+                default:
+                    return null;
             }
         }
     }
