+++ src/GUI/efreveng/TestFiles/edmx.config.json
@@ -1,0 +1,107 @@
+﻿{
+  "ConnectionString": "/Users/omatrot/Projects/EFCorePowerTools/src/GUI/NUnitTestCore/AdventureWorks2019.edmx",
+  "CodeGenerationMode": 1,
+  "ContextClassName": "Adventureworkslt2019Context",
+  "ContextNamespace": "EdmxContext",
+  "DefaultDacpacSchema": null,
+  "FilterSchemas": false,
+  "IncludeConnectionString": false,
+  "ModelNamespace": "EdmxModel",
+  "OutputContextPath": null,
+  "OutputPath": "/Users/omatrot/Projects/TestEdmx",
+  "ProjectPath": "",
+  "ProjectRootNamespace": "ConsoleApp1",
+  "Schemas": null,
+  "SelectedHandlebarsLanguage": 0,
+  "SelectedToBeGenerated": 0,
+  "Tables": [
+    {
+      "Name": "[dbo].[BuildVersion]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[dbo].[ErrorLog]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[SalesLT].[Address]",
+      "ObjectType": 0
+    },
+    {
+      "ExcludedColumns": [
+        "MiddleName"
+      ],
+      "Name": "[SalesLT].[Customer]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[SalesLT].[CustomerAddress]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[SalesLT].[Product]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[SalesLT].[ProductCategory]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[SalesLT].[ProductDescription]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[SalesLT].[ProductModel]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[SalesLT].[ProductModelProductDescription]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[SalesLT].[SalesOrderDetail]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[SalesLT].[SalesOrderHeader]",
+      "ObjectType": 0
+    },
+    {
+      "Name": "[SalesLT].[vGetAllCategories]",
+      "ObjectType": 3
+    },
+    {
+      "Name": "[SalesLT].[vProductAndDescription]",
+      "ObjectType": 3
+    },
+    {
+      "Name": "[SalesLT].[vProductModelCatalogDescription]",
+      "ObjectType": 3
+    },
+    {
+      "Name": "[dbo].[uspLogError]",
+      "ObjectType": 1
+    },
+    {
+      "Name": "[dbo].[uspPrintError]",
+      "ObjectType": 1
+    },
+    {
+      "Name": "[dbo].[ufnGetSalesOrderStatusText]",
+      "ObjectType": 2
+    }
+  ],
+  "UiHint": "sql1.AdventureWorksLT2019",
+  "UseBoolPropertiesWithoutDefaultSql": false,
+  "UseDatabaseNames": false,
+  "UseDbContextSplitting": false,
+  "UseFluentApiOnly": true,
+  "UseHandleBars": false,
+  "UseInflector": true,
+  "UseLegacyPluralizer": false,
+  "UseNoConstructor": false,
+  "UseNoNavigations": false,
+  "UseNodaTime": false,
+  "UseNullableReferences": false,
+  "UseSpatial": false
+}
